requirements.txt
absl-py==1.4.0
asttokens==2.0.5
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.4
conda==22.11.1
conda-build==3.23.3
conda-package-handling==1.9.0
cryptography==38.0.1
decorator==5.1.1
dnspython==2.2.1
exceptiongroup==1.0.4
executing==0.8.3
expecttest==0.1.4
filelock==3.6.0
flit_core==3.6.0
future==0.18.2
glob2==0.7
google-auth==2.17.1
google-auth-oauthlib==1.0.0
grpcio==1.53.0
hypothesis==6.61.0
idna==3.4
imageio==2.27.0
ipython==8.7.0
jedi==0.18.1
Jinja2==2.11.3
joblib==1.2.0
lazy_loader==0.2
libarchive-c==2.9
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
networkx==3.1
numpy==1.22.3
oauthlib==3.2.2
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
pkginfo==1.8.3
pluggy==1.0.0
prompt-toolkit==3.0.20
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.4
pycparser==2.21
Pygments==2.11.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-etcd==0.4.5
pytz==2022.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
setuptools==65.5.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
sympy==1.11.1
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
threadpoolctl==3.1.0
tifffile==2023.3.21
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchelastic==0.2.2
torchtext==0.14.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.7.1
types-dataclasses==0.6.6
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.3
wheel==0.37.1
